feat(added): addded base to API and Updated angular
