+++ firebase.json
@@ -8,7 +8,8 @@
         ],
         "rewrites": [{
             "source": "**",
-            "destination": "/index.html"
+            "destination": "/index.html",
+            "function": "app"
         }]
     },
     "firestore": {
