+++ functions/index.js
@@ -9,18 +9,19 @@
 const admin = require("firebase-admin");
 admin.initializeApp();
 
-const createNewTaskAPI = require("./createNewTaskAPI");
-const startNewSprintAPI = require("./startNewSprintAPI");
-const logWorkAPI = require("./logWorkAPI");
-const editPageTaskAPI = require("./editPageTaskAPI");
-const deleteTaskAPI = require("./deleteTaskAPI");
-const createNewUserAPI = require("./createNewUserAPI");
-const updateSprintStatusAPI = require("./updateSprintStatusAPI");
-const createNewOrgAPI = require("./createNewOrgAPI");
-const patch1API = require("./patch1API");
-const verifyUserAPI = require("./verifyUserAPI");
-const updateUserProfileAPI = require("./updateUserProfileAPI");
-const updateThemeAPI = require("./updateUserProfileAPI");
+// const createNewTaskAPI = require("./createNewTaskAPI");
+// const startNewSprintAPI = require("./startNewSprintAPI");
+// const logWorkAPI = require("./logWorkAPI");
+// const editPageTaskAPI = require("./editPageTaskAPI");
+// const deleteTaskAPI = require("./deleteTaskAPI");
+// const createNewUserAPI = require("./createNewUserAPI");
+const { users } = require("./model/users/users");
+// const updateSprintStatusAPI = require("./updateSprintStatusAPI");
+// const createNewOrgAPI = require("./createNewOrgAPI");
+// const patch1API = require("./patch1API");
+// const verifyUserAPI = require("./verifyUserAPI");
+// const updateUserProfileAPI = require("./updateUserProfileAPI");
+// const updateThemeAPI = require("./updateUserProfileAPI");
 
 
 // class Task {
@@ -43,16 +44,16 @@ const updateThemeAPI = require("./updateUserProfileAPI");
 //     }
 // }
 
-exports.createNewTask = createNewTaskAPI.createNewTask;
-exports.startNewSprint = startNewSprintAPI.startNewSprint;
-exports.logWork = logWorkAPI.logWork;
-exports.editPageTask = editPageTaskAPI.editPageTask;
-exports.deleteTask = deleteTaskAPI.deleteTask;
-exports.createNewUser = createNewUserAPI.createNewUser;
-exports.updateSprintStatus = updateSprintStatusAPI.updateSprintStatus;
-exports.createNewOrganization = createNewOrgAPI.createNewOrganization;
-exports.createNewTeamWithLabels = createNewOrgAPI.createNewTeamWithLabels;
-exports.patch1 = patch1API.patch1;
-exports.verifyUser = verifyUserAPI.verifyUser;
-exports.updateUserProfile = updateUserProfileAPI.updateUserProfile;
-exports.updateTheme = updateThemeAPI.updateTheme;
+// exports.createNewTask = createNewTaskAPI.createNewTask;
+// exports.startNewSprint = startNewSprintAPI.startNewSprint;
+// exports.logWork = logWorkAPI.logWork;
+// exports.editPageTask = editPageTaskAPI.editPageTask;
+// exports.deleteTask = deleteTaskAPI.deleteTask;
+exports.users = users;
+// exports.updateSprintStatus = updateSprintStatusAPI.updateSprintStatus;
+// exports.createNewOrganization = createNewOrgAPI.createNewOrganization;
+// exports.createNewTeamWithLabels = createNewOrgAPI.createNewTeamWithLabels;
+// exports.patch1 = patch1API.patch1;
+// exports.verifyUser = verifyUserAPI.verifyUser;
+// exports.updateUserProfile = updateUserProfileAPI.updateUserProfile;
+// exports.updateTheme = updateThemeAPI.updateTheme;
