+++ functions/model/users/lib.js
@@ -1,0 +1,24 @@
+/* eslint-disable object-curly-spacing */
+/* eslint-disable eol-last */
+/* eslint-disable indent */
+/* eslint-disable max-len */
+// const admin = require("firebase-admin");
+// const db = admin.firestore();
+
+const { db } = require("../application/lib");
+
+exports.setUser = function(Uid, PhotoURL, DisplayName, Email, PhoneNumber, ProviderId, AppKey = "", TeamId = "", AboutMe = "", AppTheme = "theme-light") {
+    const userData = db.collection("Users").doc(Uid).set({
+        AppKey: AppKey,
+        TeamId: TeamId,
+        uid: Uid,
+        photoURL: PhotoURL,
+        displayName: DisplayName,
+        email: Email,
+        phoneNumber: PhoneNumber,
+        providerId: ProviderId,
+        AboutMe: AboutMe,
+        AppTheme: AppTheme,
+    });
+    return Promise.resolve(userData);
+};
