+++ functions/model/users/users.js
@@ -1,0 +1,18 @@
+/* eslint-disable object-curly-spacing */
+/* eslint-disable eol-last */
+/* eslint-disable indent */
+/* eslint-disable max-len */
+// eslint-disable-next-line no-dupe-else-if
+
+const { functions, cors } = require("../application/lib");
+const { createNewUser } = require("../users/createNewUser");
+
+exports.users = functions.https.onRequest((request, response) => {
+    cors(request, response, () => {
+        const mode = request.body.data.mode;
+
+        if (mode == "create") {
+            return createNewUser(request, response);
+        }
+    });
+});
