+++ functions/package.json
@@ -15,6 +15,7 @@
   "main": "index.js",
   "dependencies": {
     "cors": "^2.8.5",
+    "express": "^4.17.1",
     "firebase-admin": "^9.2.0",
     "firebase-functions": "^3.11.0"
   },
