+++ package.json
@@ -1,65 +1,65 @@
 {
-  "name": "worktrolly",
-  "version": "0.2.0",
-  "scripts": {
-    "ng": "ng",
-    "start": "ng serve",
-    "build": "ng build",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
-  },
-  "private": true,
-  "dependencies": {
-    "@angular/animations": "~12.1.0",
-    "@angular/common": "~12.1.0",
-    "@angular/compiler": "~12.1.0",
-    "@angular/core": "~12.1.0",
-    "@angular/fire": "^6.1.4",
-    "@angular/forms": "~12.1.0",
-    "@angular/platform-browser": "~12.1.0",
-    "@angular/platform-browser-dynamic": "~12.1.0",
-    "@angular/router": "~12.1.0",
-    "angular-google-charts": "^2.0.0",
-    "bootstrap": "^4.5.3",
-    "cors": "^2.8.5",
-    "firebase": "^7.0.0 || ^8.0.0",
-    "firebase-functions": "^3.13.2",
-    "jquery": "^3.5.1",
-    "latest": "^0.2.0",
-    "popper.js": "^1.16.1",
-    "rxjs": "~6.6.0",
-    "stream": "0.0.2",
-    "timers": "^0.1.1",
-    "tslib": "^2.0.0",
-    "xml2js": "^0.4.23",
-    "zone.js": "~0.11.4"
-  },
-  "devDependencies": {
-    "@angular-devkit/architect": "0.1201.0",
-    "@angular-devkit/build-angular": "~0.1002.0",
-    "@angular/cli": "^12.1.0",
-    "@angular/compiler-cli": "^12.1.0",
-    "@types/jasmine": "~3.5.0",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "^12.11.1",
-    "@types/xml2js": "^0.4.8",
-    "codelyzer": "^6.0.0",
-    "firebase-tools": "^9.2.2",
-    "fuzzy": "^0.1.3",
-    "inquirer": "^6.2.2",
-    "inquirer-autocomplete-prompt": "^1.0.1",
-    "jasmine-core": "~3.6.0",
-    "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~5.0.0",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~3.0.2",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
-    "open": "^7.0.3",
-    "protractor": "~7.0.0",
-    "ts-node": "~8.3.0",
-    "tslint": "~6.1.0",
-    "typescript": "~4.3.4"
-  }
+    "name": "worktrolly",
+    "version": "0.2.0",
+    "scripts": {
+        "ng": "ng",
+        "start": "ng serve",
+        "build": "ng build",
+        "test": "ng test",
+        "lint": "ng lint",
+        "e2e": "ng e2e"
+    },
+    "private": true,
+    "dependencies": {
+        "@angular/animations": "~12.1.0",
+        "@angular/common": "~12.1.0",
+        "@angular/compiler": "~12.1.0",
+        "@angular/core": "~12.1.0",
+        "@angular/fire": "^6.1.4",
+        "@angular/forms": "~12.1.0",
+        "@angular/platform-browser": "~12.1.0",
+        "@angular/platform-browser-dynamic": "~12.1.0",
+        "@angular/router": "~12.1.0",
+        "angular-google-charts": "^2.0.0",
+        "bootstrap": "^4.5.3",
+        "cors": "^2.8.5",
+        "firebase": "^7.0.0 || ^8.0.0",
+        "firebase-functions": "^3.13.2",
+        "jquery": "^3.5.1",
+        "latest": "^0.2.0",
+        "popper.js": "^1.16.1",
+        "rxjs": "~6.6.0",
+        "stream": "0.0.2",
+        "timers": "^0.1.1",
+        "tslib": "^2.0.0",
+        "xml2js": "^0.4.23",
+        "zone.js": "~0.11.4"
+    },
+    "devDependencies": {
+        "@angular-devkit/architect": "0.1201.0",
+        "@angular-devkit/build-angular": "~0.1002.0",
+        "@angular/cli": "^12.1.0",
+        "@angular/compiler-cli": "^12.1.0",
+        "@types/jasmine": "~3.5.0",
+        "@types/jasminewd2": "~2.0.3",
+        "@types/node": "^12.11.1",
+        "@types/xml2js": "^0.4.8",
+        "codelyzer": "^6.0.0",
+        "firebase-tools": "^9.2.2",
+        "fuzzy": "^0.1.3",
+        "inquirer": "^6.2.2",
+        "inquirer-autocomplete-prompt": "^1.0.1",
+        "jasmine-core": "~3.6.0",
+        "jasmine-spec-reporter": "~5.0.0",
+        "karma": "~5.0.0",
+        "karma-chrome-launcher": "~3.1.0",
+        "karma-coverage-istanbul-reporter": "~3.0.2",
+        "karma-jasmine": "~4.0.0",
+        "karma-jasmine-html-reporter": "^1.5.0",
+        "open": "^7.0.3",
+        "protractor": "~7.0.0",
+        "ts-node": "~8.3.0",
+        "tslint": "~6.1.0",
+        "typescript": "~4.3.4"
+    }
 }
